# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDesk Remote API
  version: 0.1.0
  description: REST API skeleton for OpenDesk Remote control plane.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed session bearer token backed by a server-side session hash, or user-bound OpenDesk API token.
    sessionCookie:
      type: apiKey
      in: cookie
      name: opendesk_session
paths:
...
  /api/v1/users:
    get:
      summary: List users
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Users returned }
        "401": { description: Authentication required }
    post:
      summary: Create user
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "201": { description: User created }
        "401": { description: Authentication required }
  /api/v1/user-groups:
    get:
      summary: List user groups
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: User groups returned }
        "401": { description: Authentication required }
    post:
      summary: Create user group
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                member_user_ids:
                  type: array
                  items: { type: integer }
      responses:
        "201": { description: User group created }
        "400": { description: Invalid user group }
        "401": { description: Authentication required }
...
  /api/v1/users/{id}:
    get:
      summary: Get user
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: User returned }
        "401": { description: Authentication required }
        "404": { description: User not found }
    put:
      summary: Update user
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username]
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                display_name:
                  type: string
                status:
                  type: string
                  enum: [active, disabled, locked]
                mfa_enabled:
                  type: boolean
                source:
                  type: string
                  enum: [local, oidc, ldap]
      responses:
        "200": { description: User updated }
        "400": { description: Invalid user }
        "401": { description: Authentication required }
        "404": { description: User not found }
    delete:
      summary: Delete or disable user
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: User deleted or disabled }
        "401": { description: Authentication required }
        "404": { description: User not found }
...